Sets the base for the monitoringAPI with the interceptor pettern
